backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
bcrypt<4.2.0  # Pin to <4.2.0 for passlib compatibility
openai>=1.0.0  # AsyncOpenAI client for Gemini endpoint
openai-agents>=0.1.0  # OpenAI Agents SDK for agent orchestration
fastmcp>=0.1.0  # Official MCP SDK for tool definitions
tenacity>=8.0.0  # Retry logic with exponential backoff
alembic>=1.13.0  # Database migrations
